Xcode 3.2.6 warnings squashed;

+++ zenilib/include/Zeni/Matrix4f.hxx
@@ -103,7 +103,7 @@ namespace Zeni {
 
   Matrix4f Matrix4f::Perspective(const float &fov_rad_y, const float &aspect, const float &near_, const float &far_) {
 #define cotangent(angle) (1.0f / tan(angle))
-    const float f = cotangent(0.5f * fov_rad_y);
+    const float f = float(cotangent(0.5f * fov_rad_y));
 #undef cotangent
     const float denom = near_ - far_;
