+++ zenilib/include/Zeni/Vector3f.hxx
@@ -140,7 +140,7 @@ namespace Zeni {
   }
 
   float Vector3f::magnitude() const {
-    return sqrt(magnitude2());
+    return float(sqrt(magnitude2()));
   }
 
   Vector3f Vector3f::get_i() const {
@@ -180,7 +180,7 @@ namespace Zeni {
     const float b = rhs.magnitude();
     const float c = (rhs - *this).magnitude();
 
-    return acos((a * a + b * b - c * c) / (2 * a * b));
+    return float(acos((a * a + b * b - c * c) / (2 * a * b)));
   }
   
   const float & Vector3f::operator[](const int &index) const {
